requirements.txt
asgiref==3.7.2
black==23.3.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-geolocation-fields==0.1.8
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
